Начал переделывать алгоритм Z-буфера

+++ Lab6/AffineTransformations3D/AffineTransformations3D/Point3D.cs
@@ -5,7 +5,7 @@ namespace AffineTransformations3D
 {
     public class Point3D : IIdentifiable<long>, ICloneable
     {
-        private static int nextIdentifier = 0;
+        private static long nextIdentifier = 0;
         public double X { get; set; }
         public double Y { get; set; }
         public double Z { get; set; }
